index: add device lookup method based on database
When the service is open on a port. It will try to fetch
devices from the database and store them locally. So when
the gateway forward the message from device to the network
server, it will read the address included in the message
and compare with the local storage (a Map). If the address
exist then we get its NWSKEY and APPSKEY and ready for
asconmac decryption.