index: add method to store device metadata to db
This commit support storing device metadata to the database.
This could be useful where the Application server need to query
for these info and display them for the user. Right now, this
commit support field package_count - which store the number of
package that a particular device has sent to the network server
in a day. This will be refresh each day (creating a new collection).