index: add reliability test method
Admin are provided with a web page to start the test.
Once the test is started then the Network Server
captures incoming message in test format from node.
Admin can end the test at any time, the Network
server will process the captured message and defer
it to a python script to draw a plot.