add join-request support
Allow admin to add AppEUI, DevEUI and AppKey (device provision). The
commit provide a simple UI to let user add these info.

Refactor how sensorDevCollection layout its document. Since with
OTAA, the network server provide the device with the address. So
we use DevEUI to identify devices (as it's a unique global identifier)
Fix the local storage of sensorDevCollection too (var devicesInfo).

The join-request use the info in the device provision step to check
the MIC and authenticate the message.